Added ability to export multiple accounts into one file.

+++ AccountManager.Core/Models/RiotGames/League/Ranked/RankedTeamFightTacticsFlex.cs
@@ -5,12 +5,12 @@ namespace AccountManager.Core.Models.RiotGames.League
     public sealed class RankedTeamFightTacticsFlex
     {
         [JsonPropertyName("currentSeasonId")]
-        public int CurrentSeasonId { get; set; }
+        public long CurrentSeasonId { get; set; }
 
         [JsonPropertyName("currentSeasonEnd")]
         public long CurrentSeasonEnd { get; set; }
 
         [JsonPropertyName("nextSeasonStart")]
-        public int NextSeasonStart { get; set; }
+        public long NextSeasonStart { get; set; }
     }
 }
